CrackYourInterview/requirements.txt
fastapi
uvicorn
pdfminer.six
python-docx
python-multipart
